Renames sendTelegramDidRewardCall to sendDelegatorTelegram Renames sendTelegramClaimRewardCall to sendDelegateTelegram Also fixs parameters usage
